+++ CHANGELOG.md
@@ -6,9 +6,10 @@ All notable changes to the "playwright-helpers" extension will be documented in 
 
 - TBD
 
-## [1.2.28] - 2026-03-04
+## [1.2.28] - 2026-03-05
 
-- feat: new Playwright CLI command
+- feat: all Playwright CLI commands in new category "Playwright CLI"
+- fix: Playwright CLI command for listing package version
 
 ## [1.2.27] - 2026-03-03